moved all scripts to scripts folder
