+++ scripts/ghostBuster.js
@@ -1,8 +1,8 @@
 //right now, to add or edit team information, you'll use a file named teams.js in the root of the project
 //TODO: (maybe? not sure if it's worth the effort) integrate with Google Sheets to automate the team information
 
-const { thesisTeams, greenfieldTeams, legacyTeams } = require('./config/teams');
-const Team = require('./helpers/team');
+const { thesisTeams, greenfieldTeams, legacyTeams } = require('../config/teams');
+const Team = require('../helpers/team');
 
 let ghostMessages = [];
 let potentialGhostMessages = [];
